cloudtasks: refactor from CLI to snippets.
