+++ cloudtasks/pull_queue/README.md
@@ -29,50 +29,14 @@ To create a queue using the Cloud SDK, use the following gcloud command:
 
     gcloud beta tasks queues create-pull-queue my-pull-queue
 
-## Running the Samples
-
-Set the environment variables:
-
-First, your project ID:
-
-    export PROJECT_ID=my-project-id
-
-Then the queue ID, as specified at queue creation time. Queue IDs already
-created can be listed with `gcloud beta tasks queues list`.
-
-    export QUEUE_ID=my-pull-queue
-
-And finally the location ID, which can be discovered with
-`gcloud beta tasks queues describe $QUEUE_ID`, with the location embedded in
-the "name" value (for instance, if the name is
-"projects/my-project/locations/us-central1/queues/my-pull-queue", then the
-location is "us-central1").
-
-    export LOCATION_ID=us-central1
-
-## Sample CLI Operations
-
-Create a task for a queue:
-
-    go run main.go create $PROJECT_ID $LOCATION_ID $QUEUE_ID
-
-Pull a task:
-
-    go run main.go pull $PROJECT_ID $LOCATION_ID $QUEUE_ID
-
-Acknowledge task:
-
-    go run main.go acknowledge <task>
-
-* where task is the output from pull task, example:  
-`'{"name":"projects/my-project-id/locations/us-central1/queues/my-queue/tasks/1234","scheduleTime":"2017-11-01T22:27:
-  53.628279Z"}'`
-
 ## Runnning the Tests
 
-Use the built-in environment variable to set the Project ID.
+1. Enable the Cloud Tasks API for your project
+2. Pre-create a Cloud Tasks Pull Queue
+3. Set the project ID environment variable:
 
 ```
 export GOLANG_SAMPLES_PROJECT_ID=my-project-id
 ```
 
+With those steps in place, `go test -v ./...`.
