+++ cloudtasks/pull_queue/lease_task_test.go
@@ -2,46 +2,32 @@
 // Use of this source code is governed by the Apache 2.0
 // license that can be found in the LICENSE file.
 
-package main
+package snippets
 
 import (
+	"io/ioutil"
 	"testing"
 
 	"github.com/GoogleCloudPlatform/golang-samples/internal/testutil"
 )
 
-var projectID, locationID, queueID string
-
-func setup(t *testing.T) {
+func TestTaskLeaseAndAck(t *testing.T) {
 	tc := testutil.SystemTest(t)
 
 	// ProjectID is set from environment variable GOLANG_SAMPLES_PROJECT_ID.
-	projectID = tc.ProjectID
-	locationID = "us-central1"
-	queueID = "my-pull-queue"
-}
-
-func TestTaskCreate(t *testing.T) {
-	setup(t)
-
-	_, err := taskCreate(projectID, locationID, queueID)
-	if err != nil {
-		t.Fatalf("failed to create new task: %v", err)
-	}
-}
-
-func TestTaskLeaseAndAck(t *testing.T) {
-	setup(t)
+	projectID := tc.ProjectID
+	locationID := "us-central1"
+	queueID := "my-pull-queue"
 
 	// Guarantee a task will be available in the queue
 	// in the event TestTaskCreate is skipped.
-	_, err := taskCreate(projectID, locationID, queueID)
+	seed, err := taskCreate(ioutil.Discard, projectID, locationID, queueID)
 	if err != nil {
 		t.Fatalf("failed to ensure a task would be available for lease: %v", err)
 	}
 
 	// Test task leasing.
-	task, err := taskLease(projectID, locationID, queueID)
+	task, err := taskLease(ioutil.Discard, projectID, locationID, queueID)
 	if err != nil {
 		t.Fatalf("failed to lease a task: %v", err)
 	}
@@ -49,8 +35,13 @@ func TestTaskLeaseAndAck(t *testing.T) {
 		t.Fatalf("no task available to lease: %v", err)
 	}
 
+	// Note cross-test data poisoning through concurrent queue usage or pre-created tasks.
+	if task.GetName() != seed.GetName() {
+		t.Logf("Task used for lease testing was not created by test setup: (seeded: %s, leased: %s)", seed.GetName(), task.GetName())
+	}
+
 	// Acknowledge our leased task.
-	if err = taskAck(task); err != nil {
+	if err = taskAck(ioutil.Discard, task); err != nil {
 		t.Fatalf("failed to acknowledge task: %v", err)
 	}
 }
